Library

Library

Directed acyclic graphs (DAGs)

CausalInference.dsep — Function.

dsep(g::AbstractGraph, u, v, s; verbose = false)

Check whether u and v are d-separated given set s. Algorithm: unrolled https://arxiv.org/abs/1304.1505

source

CausalInference.cpdag — Function.

cpdag(skel::DiGraph)

Reference: M. Chickering: Learning equivalence classes of Bayesian network structures. Journal of Machine Learning Research 2 (2002). M. Chickering: A Transformational Characterization of Equivalent Bayesian Network Structures. (1995).

Note that the edge order defined there is already partly encoded into the representation of a DiGraph.

source

CausalInference.vskel — Function.

vskel(g)

Skeleton and v-structures. (Currently from the first step of the pc-Alg.)

source

PC algorithm

CausalInference.pcalg — Function.

pcalg(n::V, I, par...)

Perform the PC skeleton algorithm for a set of 1:n variables using the tests

I(u, v, [s1, ..., sn], par...)

Returns the CPDAG as DiGraph.

source

CausalInference.skeleton — Function.

skeleton(n, I) -> g, S

Perform the undirected PC skeleton algorithm for a set of 1:n variables using the test I. Returns skeleton graph and separating set

source

CausalInference.dseporacle — Function.

dseporacle(i, j, s, g)

Oracle for the skeleton and pcalg functions using dsep on the true graph g

source

CausalInference.unshielded — Function.

unshielded(g, S)

Find unshielded triples in the skeleton. Triples are connected vertices v-w-z where z is not a neighbour of v. Uses that edges iterates in lexicographical order.

source

Statistics

CausalInference.gausscitest — Function.

gausscitest(i, j, s, (C,n), c)

Test for conditional independence of variable no i and j given variables in s with Gaussian test at the critical value c. C is covariance of n observations.

source

CausalInference.partialcor — Function.

partialcor(i, j, s, C)

Compute the partial correlation of nodes i and j given list of nodes s using the correlation matrix C.

source

Miscellaneous

CausalInference.digraph — Function.

digraph(E)

Create DiGraph from edge-list.

source

CausalInference.pairs — Function.

pairs(g)

Return the edge-list as Pairs.

source

CausalInference.pc_oracle — Function.

pc_oracle(g)

Compute CPDAG using the PC algorithm using the dseporacle on the DAG g.

source

CausalInference.skel_oracle — Function.

skel_oracle(g)

Compute the skeleton using the dseporacle for the DAG g.

source

CausalInference.randdag — Function.

randdag(n, alpha = 0.1)

Create random DAG from randomly permuted random triangular matrix with edge probability alpha.

source

CausalInference.disjoint_sorted — Function.

disjoint_sorted(u, v)

Check if the intersection of sorted collections is empty. The intersection of empty collectios is empty.

source

CausalInference.ordered_edges — Function.

ordered_edges(dag)

Iterator of edges of a dag, ordered in Chickering order:

Perform a topological sort on the NODES
while there are unordered EDGES in g
    Let y be the lowest ordered NODE that has an unordered EDGE incident into it
    Let x be the highest ordered NODE for which x => y is not ordered
    return x => y 
end

source

CausalInference.insorted — Function.

insorted(a, x)

Check if x is in the sorted collection a

source

CausalInference.removesorted! — Function.

removesorted(collection, item) -> contains(collection, item)

Remove item from sorted collection.

source