Library

Library

Directed acyclic graphs (DAGs)

CausalInference.dsep — Function.

dsep(g::AbstractGraph, u, v, s; verbose = false)

Check whether u and v are d-separated given set s. Algorithm: unrolled https://arxiv.org/abs/1304.1505

source

CausalInference.cpdag — Function.

cpdag(skel::DiGraph)

Reference: M. Chickering: Learning equivalence classes of Bayesian network structures. Journal of Machine Learning Research 2 (2002). M. Chickering: A Transformational Characterization of Equivalent Bayesian Network Structures. (1995).

Note that the edge order defined there is already partly encoded into the representation of a DiGraph.

source

CausalInference.vskel — Function.

vskel(g)

Skeleton and v-structures. (Currently from the first step of the pc-Alg.)

source

PC algorithm

CausalInference.pcalg — Function.

pcalg(n::V, I, par...)

Perform the PC algorithm for a set of 1:n variables using the tests

I(u, v, [s1, ..., sn], par...)

Returns the CPDAG as DiGraph.

source

pcalg(t, p::Float64, test::typeof(gausscitest); kwargs...)

run PC algorithm for tabular input data t using a p-value p to test for conditional independeces using Fisher's z-transformation.

source

pcalg(t::T, p::Float64; cmitest::typeof(cmitest); kwargs...) where{T}

run PC algorithm for tabular input data t using a p-value p to detect conditional independeces using a conditional mutual information permutation test.

source

CausalInference.skeleton — Function.

skeleton(n, I) -> g, S

Perform the undirected PC skeleton algorithm for a set of 1:n variables using the test I. Returns skeleton graph and separating set

source

CausalInference.dseporacle — Function.

dseporacle(i, j, s, g)

Oracle for the skeleton and pcalg functions using dsep on the true graph g

source

CausalInference.unshielded — Function.

unshielded(g, S)

Find unshielded triples in the skeleton. Triples are connected vertices v-w-z where z is not a neighbour of v. Uses that edges iterates in lexicographical order.

source

Statistics

CausalInference.gausscitest — Function.

gausscitest(i, j, s, (C,n), c)

Test for conditional independence of variable no i and j given variables in s with Gaussian test at the critical value c. C is covariance of n observations.

source

CausalInference.partialcor — Function.

partialcor(i, j, s, C)

Compute the partial correlation of nodes i and j given list of nodes s using the correlation matrix C.

source

CausalInference.cmitest — Function.

cmitest(i,j,s,data,crit; kwargs...)

Test for conditional independence of variables i and j given variables in s with permutation test using nearest neighbor conditional mutual information estimates at p-value crit.

keyword arguments: kwargs...: keyword arguments passed to independence tests

source

KL Entropy Estimators

CausalInference.n_ball — Function.

n_ball(n::Number)

Computes the volume of a n-dimensional unit sphere.

source

CausalInference.kl_entropy — Function.

kl_entropy(data::Array{Float64, 2}; k=5)

Compute the nearest-neighbor estimate of the differential entropy of data.

data is a 2d array, with every column representing one data point. For further information, see

"A class of Rényi information estimators for multidimensional densities" Nikolai Leonenko, Luc Pronzato, and Vippal Savani The Annals of Statistics, 2008 https://projecteuclid.org/euclid.aos/1223908088

keyword arguments: k=5: number of nearest neighbors

source

CausalInference.kl_renyi — Function.

kl_renyi(data::Array{Float64, 2}, q; k=5)

Compute the nearest-neighbor estimate of the Renyi-alpha entropy of data.

data is a 2d array, with every column representing one data point. For further information, see

"A class of Rényi information estimators for multidimensional densities" Nikolai Leonenko, Luc Pronzato, and Vippal Savani The Annals of Statistics, 2008 https://projecteuclid.org/euclid.aos/1223908088

keyword arguments: k=5: number of nearest neighbors

source

CausalInference.kl_mutual_information — Function.

kl_mutual_information(x, y; k=5, bias_correction=true)

compute the nearest-neighbor 'KGS' estimate of the mutual information between x and y.

x and y are 2d arrays, with every column representing one data point. For further information, see

"Estimating Mutual Information" Alexander Kraskov, Harald Stoegbauer, and Peter Grassberger Physical Review E https://arxiv.org/pdf/cond-mat/0305641.pdf

"Demystifying Fixed k-Nearest Neighbor Information Estimators" Weihao Gao, Sewoong Oh, Pramod Viswanath EEE International Symposium on Information Theory - Proceedings https://arxiv.org/pdf/1604.03006.pdf

keyword arguments: k=5: number of nearest neighbors bias_correction=true: flag to apply Gao's bias correction

source

CausalInference.kl_cond_mi — Function.

kl_cond_mi(x, y, z; k=5, bias_correction=true)

compute the nearest-neighbor 'KGS' estimate of the conditional mutual information between x and y given z.

x, y, and z are 2d arrays, with every column representing one data point. keyword arguments: k=5: number of nearest neighbors bias_correction=true: flag to apply Gao's bias correction

source

CausalInference.kl_perm_mi_test — Function.

kl_perm_mi_test(x, y; k=5, B=100, bias_correction=true)

compute permutation test of independence of x and y.

keyword arguments: k=5: number of nearest neighbors to use for mutual information estimate B=100: number of permutations bias_correction=true: flag to apply Gao's bias correction

source

CausalInference.kl_perm_cond_mi_test — Function.

kl_perm_cond_mi_test(x, y, z; k=5, B=100, kp=5, bias_correction=true)

compute permutation test of conditional independence of x and y given z.

For further information, see: "Conditional independence testing based on a nearest-neighbor estimator of conditional mutual information" Jakob Runge Proceedings of the 21st International Conference on Artificial Intelligence and Statistics (AISTATS) 2018, Lanzarote, Spain. http://proceedings.mlr.press/v84/runge18a/runge18a.pdf

keyword arguments: k=5: number of nearest neighbors to use for mutual information estimate B=100: number of permutations bias_correction=true: flag to apply Gao's bias correction

source

Miscellaneous

CausalInference.digraph — Function.

digraph(E)

Create DiGraph from edge-list.

source

CausalInference.vpairs — Function.

vpairs(g)

Return the edge-list as Pairs.

source

CausalInference.pc_oracle — Function.

pc_oracle(g)

Compute CPDAG using the PC algorithm using the dseporacle on the DAG g.

source

CausalInference.skel_oracle — Function.

skel_oracle(g)

Compute the skeleton using the dseporacle for the DAG g.

source

CausalInference.randdag — Function.

randdag(n, alpha = 0.1)

Create random DAG from randomly permuted random triangular matrix with edge probability alpha.

source

CausalInference.disjoint_sorted — Function.

disjoint_sorted(u, v)

Check if the intersection of sorted collections is empty. The intersection of empty collectios is empty.

source

CausalInference.ordered_edges — Function.

ordered_edges(dag)

Iterator of edges of a dag, ordered in Chickering order:

Perform a topological sort on the NODES
while there are unordered EDGES in g
    Let y be the lowest ordered NODE that has an unordered EDGE incident into it
    Let x be the highest ordered NODE for which x => y is not ordered
    return x => y 
end

source

CausalInference.insorted — Function.

insorted(a, x)

Check if x is in the sorted collection a

source

CausalInference.removesorted! — Function.

removesorted(collection, item) -> contains(collection, item)

Remove item from sorted collection.

source